Update README.md
I didn't get the difference between the labor needed to add "official" JD info and the labor needed to swap the JD info from Twitter with the "official" source.
But it should prevent users from getting misled in the future. So It seems better.
Anyway, I will do anything to get this PR merged ASAP.